feat(sdk): complete 10-language SDK — Phases 3-6 (Go, Kotlin JVM, Swift, Rust, C, HarmonyOS/ArkTS, Android, CI/CD)
Ports all 8 portable-core modules (models, memory, identity, languages, companion,
inference, tools, sync) to 7 additional languages, completing the cross-platform SDK.

Additions per language:
- Go: BiometricMatcher (float64 accumulators), FaceAffectMapper/Bridge, Goal.AdvanceProgress,
  FacialMetricMatrix + FaceExpressionClassification; new biometric + goal-progress fixture tests
- Kotlin JVM: BiometricMatcher (Double accumulators), FaceAffectMapper, FaceCompanionBridge
  (CONFUSION_THRESHOLD=0.70, expression guard), Goal.advanceProgress; new test files
- Swift: BiometricMatcher (Double, no vDSP/Accelerate), FaceAffectMapper/Bridge,
  Goal.advancingProgress, FacialMetricMatrix; new BiometricMatcherTests + GoalProgressTests
- Rust: restructured from flat .rs files to proper subdirectory modules (memory/, identity/,
  languages/, companion/, tools/); cosine_similarity using f64; FaceCompanionBridge with
  expression match guard; new test files
- C: double-precision cosineSimilarity, ca_face_apply_affect (5 expressions, conf<0.5 gate),
  ca_face_observe with is_confusion_expr guard + canonical trigger name; new test files
- HarmonyOS/ArkTS: all 8 modules as .ets files, numeric enums, scalar number[] cosine loop,
  faceObserve with Confused|Stressed guard; 5 fixture test suites
- Android: migrated to com.bhengubv.circleai.android sub-package (minSdk 26), Double
  accumulators (no NDK), FaceCompanionBridge expression guard; 5 test files covering all
  fixture vectors including BiometricMatcherTest + GoalProgressTest

All implementations share identical AffectState math (tol 1e-5), cosine similarity
(tol 1e-4), and KnownLanguages registry (exact), validated by shared fixture JSON vectors.

Phase 6 — CI/CD:
- .github/workflows/sdk-tests.yml: 10 parallel jobs + all-tests-pass gate
- .github/workflows/fixture-interop.yml: cross-language AffectState/BiometricMatcher/KnownLanguages interop
- .github/workflows/publish.yml: gated publish to PyPI, npm, crates.io, GitHub Packages, SPM

[skip ci]

+++ c/tests/test_affect_state.c
@@ -1,3 +1,10 @@
+/*
+ * test_affect_state.c — AffectState cross-language fixture tests.
+ *
+ * All expected values match fixtures/affect_state.json within epsilon=1e-5.
+ * Returns 0 on all-pass, calls assert() on first failure.
+ */
+
 #include <stdio.h>
 #include <math.h>
 #include <assert.h>
@@ -5,10 +12,11 @@
 
 #define EPSILON 1e-5f
 
-static void check(const char* id, float got, float expected) {
+static void check(const char *id, float got, float expected) {
     float diff = fabsf(got - expected);
     if (diff > EPSILON) {
-        fprintf(stderr, "FAIL [%s]: got %.8f, expected %.8f (diff %.8f)\n", id, got, expected, diff);
+        fprintf(stderr, "FAIL [%s]: got %.8f, expected %.8f (diff %.8f)\n",
+                id, (double)got, (double)expected, (double)diff);
         assert(0);
     }
 }
@@ -51,9 +59,10 @@ int main(void) {
     check("neg_twice rapport",     s.rapport,     0.0f);
     check("neg_twice energy",      s.energy,      0.5f);
 
-    /* positive_then_negative
-     * Start default, positive: engagement=0.52, uncertainty=0.18, rapport=0.01
-     * Then negative: engagement=0.52-0.03=0.49, uncertainty=0.18+0.03=0.21, rapport stays 0.01
+    /*
+     * positive_then_negative
+     * positive: engagement=0.52, uncertainty=0.18, rapport=0.01
+     * negative: engagement=0.52-0.03=0.49, uncertainty=0.18+0.03=0.21
      */
     s = ca_affect_state_default();
     ca_affect_state_positive_signal(&s);
@@ -63,9 +72,10 @@ int main(void) {
     check("pos_neg rapport",     s.rapport,     0.01f);
     check("pos_neg energy",      s.energy,      0.5f);
 
-    /* negative_then_positive
-     * Start default, negative: engagement=0.47, uncertainty=0.23, rapport=0.0
-     * Then positive: engagement=0.47+0.02=0.49, uncertainty=0.23-0.02=0.21, rapport=0.0+0.01=0.01
+    /*
+     * negative_then_positive
+     * negative: engagement=0.47, uncertainty=0.23, rapport=0.0
+     * positive: engagement=0.47+0.02=0.49, uncertainty=0.23-0.02=0.21, rapport=0+0.01=0.01
      */
     s = ca_affect_state_default();
     ca_affect_state_negative_signal(&s);
@@ -75,8 +85,9 @@ int main(void) {
     check("neg_pos rapport",     s.rapport,     0.01f);
     check("neg_pos energy",      s.energy,      0.5f);
 
-    /* idle_decay_1h
-     * Input: engagement=0.8, energy=0.7, rapport=0.4 (curiosity=0.5, uncertainty=0.2)
+    /*
+     * idle_decay_1h
+     * Input: engagement=0.8, energy=0.7, rapport=0.4
      * decay = min(0.3, 1*0.02) = 0.02
      * engagement = 0.8 + (0.5-0.8)*0.02 = 0.8 - 0.006 = 0.794
      * energy     = 0.7 + (0.5-0.7)*0.02 = 0.7 - 0.004 = 0.696
@@ -84,13 +95,13 @@ int main(void) {
     s = ca_affect_state_default();
     s.engagement = 0.8f; s.energy = 0.7f; s.rapport = 0.4f;
     ca_affect_state_idle_decay(&s, 1.0f);
-    check("decay_1h engagement", s.engagement, 0.794f);
-    check("decay_1h energy",     s.energy,     0.696f);
-    check("decay_1h rapport",    s.rapport,    0.4f);
-    check("decay_1h uncertainty",s.uncertainty,0.2f);
+    check("decay_1h engagement",  s.engagement,  0.794f);
+    check("decay_1h energy",      s.energy,      0.696f);
+    check("decay_1h rapport",     s.rapport,     0.4f);
+    check("decay_1h uncertainty", s.uncertainty, 0.2f);
 
-    /* idle_decay_8h
-     * Input: engagement=0.8, energy=0.7, rapport=0.4
+    /*
+     * idle_decay_8h
      * decay = min(0.3, 8*0.02) = 0.16
      * engagement = 0.8 + (0.5-0.8)*0.16 = 0.8 - 0.048 = 0.752
      * energy     = 0.7 + (0.5-0.7)*0.16 = 0.7 - 0.032 = 0.668
@@ -98,13 +109,13 @@ int main(void) {
     s = ca_affect_state_default();
     s.engagement = 0.8f; s.energy = 0.7f; s.rapport = 0.4f;
     ca_affect_state_idle_decay(&s, 8.0f);
-    check("decay_8h engagement", s.engagement, 0.752f);
-    check("decay_8h energy",     s.energy,     0.668f);
-    check("decay_8h rapport",    s.rapport,    0.4f);
-    check("decay_8h uncertainty",s.uncertainty,0.2f);
+    check("decay_8h engagement",  s.engagement,  0.752f);
+    check("decay_8h energy",      s.energy,      0.668f);
+    check("decay_8h rapport",     s.rapport,     0.4f);
+    check("decay_8h uncertainty", s.uncertainty, 0.2f);
 
-    /* idle_decay_24h
-     * Input: engagement=0.8, energy=0.7, rapport=0.4
+    /*
+     * idle_decay_24h
      * decay = min(0.3, 24*0.02) = 0.3 (capped)
      * engagement = 0.8 + (0.5-0.8)*0.3 = 0.8 - 0.09 = 0.71
      * energy     = 0.7 + (0.5-0.7)*0.3 = 0.7 - 0.06 = 0.64
@@ -112,16 +123,17 @@ int main(void) {
     s = ca_affect_state_default();
     s.engagement = 0.8f; s.energy = 0.7f; s.rapport = 0.4f;
     ca_affect_state_idle_decay(&s, 24.0f);
-    check("decay_24h engagement", s.engagement, 0.71f);
-    check("decay_24h energy",     s.energy,     0.64f);
-    check("decay_24h rapport",    s.rapport,    0.4f);
-    check("decay_24h uncertainty",s.uncertainty,0.2f);
+    check("decay_24h engagement",  s.engagement,  0.71f);
+    check("decay_24h energy",      s.energy,      0.64f);
+    check("decay_24h rapport",     s.rapport,     0.4f);
+    check("decay_24h uncertainty", s.uncertainty, 0.2f);
 
-    /* clamp_max_positive
+    /*
+     * clamp_max_positive
      * Input: engagement=0.99, uncertainty=0.01, rapport=0.99
-     * positive: engagement = min(1.0, 0.99+0.02) = 1.0
-     *           uncertainty= max(0.0, 0.01-0.02) = 0.0
-     *           rapport    = min(1.0, 0.99+0.01) = 1.0
+     * engagement  = min(1.0, 0.99+0.02) = 1.0
+     * uncertainty = max(0.0, 0.01-0.02) = 0.0
+     * rapport     = min(1.0, 0.99+0.01) = 1.0
      */
     s = ca_affect_state_default();
     s.engagement = 0.99f; s.uncertainty = 0.01f; s.rapport = 0.99f;
@@ -131,10 +143,11 @@ int main(void) {
     check("clamp_max rapport",     s.rapport,     1.0f);
     check("clamp_max energy",      s.energy,      0.5f);
 
-    /* clamp_min_negative
+    /*
+     * clamp_min_negative
      * Input: engagement=0.01, uncertainty=0.98
-     * negative: engagement = max(0.0, 0.01-0.03) = 0.0
-     *           uncertainty= min(1.0, 0.98+0.03) = 1.0
+     * engagement  = max(0.0, 0.01-0.03) = 0.0
+     * uncertainty = min(1.0, 0.98+0.03) = 1.0
      */
     s = ca_affect_state_default();
     s.engagement = 0.01f; s.uncertainty = 0.98f;
@@ -144,8 +157,9 @@ int main(void) {
     check("clamp_min rapport",     s.rapport,     0.0f);
     check("clamp_min energy",      s.energy,      0.5f);
 
-    /* idle_decay_neutral_no_change
-     * Input: default state (engagement=0.5, energy=0.5), apply 8h decay
+    /*
+     * idle_decay_neutral_no_change
+     * engagement=0.5, energy=0.5 (already at neutral target)
      * lerpf(0.5, 0.5, t) = 0.5 regardless of t
      */
     s = ca_affect_state_default();
